remove cast

+++ src/io/reflection/app/client/page/part/MyAppsTopPanel.java
@@ -123,7 +123,7 @@ public class MyAppsTopPanel extends Composite {
 		dateSelector.addFixedRanges(dateSelectorPresetRanges);
 
 		// Reset linked account id in filter, to avoid problems after refreshing the page
-		FilterController.get().getFilter().setLinkedAccountId((long) 0);
+		FilterController.get().getFilter().setLinkedAccountId(0L);
 
 		updateFromFilter();
